refactor(build): extract 3 cohesive concerns from lib.mjs into scripts/lib/ (#510)
scripts/lib.mjs was a 3105-LOC monolith. Moved 3 cleanly-separable concerns verbatim into dedicated
modules, re-exported from lib.mjs so all 69 importers keep their paths (pure code-motion, zero
call-site churn):
- scripts/lib/endpoint-artifacts.mjs (629) — RPC/endpoint/pool/incident artifact builders
- scripts/lib/readme-links.mjs (223) — README link selection/classification (self-contained)
- scripts/lib/economics-artifacts.mjs (118) — computeMinerReadiness + buildEconomicsArtifact
lib.mjs: 3105 → 2190 LOC.

INVARIANT (gate): build artifacts byte-identical before/after — verified across 2252 artifacts
(master sha256 matches). Added tests/artifacts.test.mjs "build is deterministic (byte-identical
across rebuilds)": a data-agnostic full-tree digest guard, proven genuine (injecting Date.now() into
a moved module makes it fail). Left build-artifacts.mjs (module-level mutable state) + the
interconnected URL-safety clusters for a follow-up. 1543 tests + dry-run green.
